backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.0.0
python-dotenv==1.0.0
httpx==0.24.1
PyJWT==2.8.0